Add data.js file and modules folder
