+++ src/index.js
@@ -1,35 +1,9 @@
 import './style.css';
+import table from './modules/data';
 
 const tableDisplay = document.querySelector('.dinamic-table');
 tableDisplay.innerHTML= '<ul class="row"></ul>';
 tableDisplay.className= 'dinamicTable';
-
-const table = [
-  {
-    name: 'Mariana',
-    score: 12345
-  },
-  {
-    name: 'Fernando',
-    score: 12345
-  },
-  {
-    name: 'Daniela',
-    score: 12345
-  },
-  {
-    name: 'Julio',
-    score: 12345
-  },
-  {
-    name: 'Viviana',
-    score: 12345
-  },
-  {
-    name: 'Marcos',
-    score: 12345
-  }
-];
 
 table.forEach(user => {
   const line = document.createElement('li');
@@ -37,4 +11,4 @@ table.forEach(user => {
   line.className = 'line';
   line.classList.add('row');
   tableDisplay.appendChild(line);
-})
+});
