+++ src/modules/data.js
@@ -1,0 +1,28 @@
+const table = [
+  {
+    name: 'Mariana',
+    score: 12345
+  },
+  {
+    name: 'Fernando',
+    score: 12345
+  },
+  {
+    name: 'Daniela',
+    score: 12345
+  },
+  {
+    name: 'Julio',
+    score: 12345
+  },
+  {
+    name: 'Viviana',
+    score: 12345
+  },
+  {
+    name: 'Marcos',
+    score: 12345
+  }
+];
+
+export default table;
